refactor: move glob matching from internal/analysis to internal/index
Exported as index.MatchGlob so internal/index can use it for ADR scope
matching (added in a follow-up commit) without an import cycle.
